Update sync unit tests to build on linux Review URL: http://codereview.chromium.org/218014
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28349 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 28d0c9c14970a0e1c3c6dcc0746529446c692393

+++ chrome/browser/sync/engine/syncer_proto_util_unittest.cc
@@ -21,11 +21,11 @@ TEST(SyncerProtoUtil, TestBlobToProtocolBufferBytesUtilityFunctions) {
   unsigned char test_data3[] = {99, 2, 3, 4, 5, 6, 7, 8};
 
   syncable::Blob test_blob1, test_blob2, test_blob3;
-  for (int i = 0; i < arraysize(test_data1); ++i)
+  for (size_t i = 0; i < arraysize(test_data1); ++i)
     test_blob1.push_back(test_data1[i]);
-  for (int i = 0; i < arraysize(test_data2); ++i)
+  for (size_t i = 0; i < arraysize(test_data2); ++i)
     test_blob2.push_back(test_data2[i]);
-  for (int i = 0; i < arraysize(test_data3); ++i)
+  for (size_t i = 0; i < arraysize(test_data3); ++i)
     test_blob3.push_back(test_data3[i]);
 
   string test_message1(reinterpret_cast<char*>(test_data1),
